requirements.txt
streamlit
python-dotenv
requests
sqlite-utils
openpyxl
xgboost
joblib
numpy
scikit-learn
